refactor: move into main script and update azureconst.go
