+++ pkg/helpers/azureconst.go
@@ -1463,3 +1463,89 @@ func GetSizeMap() string {
    }
 `
 }
+
+// AcceleratedNetworkingSkus are those Azure VM SKUs that support accelerated networking.
+var AcceleratedNetworkingSkus []string = []string{
+	"AZAP_Performance_ComputeV17C",
+	"SQLGL",
+	"SQLGLCore",
+	"Standard_B12ms",
+	"Standard_B16ms",
+	"Standard_B20ms",
+	"Standard_D2_v2",
+	"Standard_D3_v2",
+	"Standard_D4_v2",
+	"Standard_D5_v2",
+	"Standard_D11_v2",
+	"Standard_D12_v2",
+	"Standard_D13_v2",
+	"Standard_D14_v2",
+	"Standard_D15_v2",
+	"Standard_F2",
+	"Standard_F4",
+	"Standard_F8",
+	"Standard_F16",
+	"Standard_DS2_v2",
+	"Standard_DS3_v2",
+	"Standard_DS4_v2",
+	"Standard_DS5_v2",
+	"Standard_DS11-1_v2",
+	"Standard_DS11_v2",
+	"Standard_DS12-1_v2",
+	"Standard_DS12-2_v2",
+	"Standard_DS12_v2",
+	"Standard_DS13-2_v2",
+	"Standard_DS13-4_v2",
+	"Standard_DS13_v2",
+	"Standard_DS14-4_v2",
+	"Standard_DS14-8_v2",
+	"Standard_DS14_v2",
+	"Standard_DS15_v2",
+	"Standard_F2s",
+	"Standard_F4s",
+	"Standard_F8s",
+	"Standard_F16s",
+	"Standard_D4_v3",
+	"Standard_D8_v3",
+	"Standard_D16_v3",
+	"Standard_D32_v3",
+	"Standard_D4s_v3",
+	"Standard_D8s_v3",
+	"Standard_D16s_v3",
+	"Standard_D32s_v3",
+	"Standard_D48_v3",
+	"Standard_D64_v3",
+	"Standard_D48s_v3",
+	"Standard_D64s_v3",
+	"Standard_E4_v3",
+	"Standard_E8_v3",
+	"Standard_E16_v3",
+	"Standard_E32_v3",
+	"Standard_E48_v3",
+	"Standard_E64i_v3",
+	"Standard_E64_v3",
+	"Standard_E4-2s_v3",
+	"Standard_E4s_v3",
+	"Standard_E8-2s_v3",
+	"Standard_E8-4s_v3",
+	"Standard_E8s_v3",
+	"Standard_E16-4s_v3",
+	"Standard_E16-8s_v3",
+	"Standard_E16s_v3",
+	"Standard_E20s_v3",
+	"Standard_E32-8s_v3",
+	"Standard_E32-16s_v3",
+	"Standard_E32s_v3",
+	"Standard_E48s_v3",
+	"Standard_E64-16s_v3",
+	"Standard_E64-32s_v3",
+	"Standard_E64is_v3",
+	"Standard_E64s_v3",
+	"Standard_F4s_v2",
+	"Standard_F8s_v2",
+	"Standard_F16s_v2",
+	"Standard_F32s_v2",
+	"Standard_F48s_v2",
+	"Standard_F64s_v2",
+	"Standard_F72s_v2",
+}
